search results, interfaces, and import clean up
