+++ cypress/component/avatar.cy.tsx
@@ -1,4 +1,4 @@
-import Avatar from "../../src/components/avatar";
+import Avatar from "@/src/components/avatar";
 
 describe("<Avatar />", () => {
   it("should render and display expected content", () => {
