+++ src/components/hero-post.tsx
@@ -1,17 +1,17 @@
-import type Author from "../interfaces/author";
+import Link from "next/link";
+import { Author } from "../interfaces/author";
 import Avatar from "./avatar";
 import CoverImage from "./cover-image";
 import DateFormatter from "./date-formatter";
-import Link from "next/link";
 
-type Props = {
+interface Props {
   title: string;
   coverImage: string;
   date: string;
   excerpt: string;
   author: Author;
   slug: string;
-};
+}
 
 const HeroPost = ({
   title,
