+++ src/components/post-footer.tsx
@@ -3,10 +3,10 @@ import { Post } from "../interfaces/post";
 import { BASE_URL } from "../lib/constants";
 import MoreStories from "./more-stories";
 
-type Props = {
+interface Props {
   morePosts: Post[];
   slug: string;
-};
+}
 
 export default function PostFooter({ morePosts, slug }: Props) {
   const shareURL = BASE_URL + slug;
