+++ src/components/post-title.tsx
@@ -1,8 +1,8 @@
 import { ReactNode } from "react";
 
-type Props = {
+interface Props {
   children?: ReactNode;
-};
+}
 
 const PostTitle = ({ children }: Props) => {
   return (
