+++ src/components/reload-page-button.tsx
@@ -3,9 +3,9 @@
 import { useRouter } from "next/navigation";
 import { ReactNode } from "react";
 
-type Props = {
+interface Props {
   children: ReactNode;
-};
+}
 
 export default function ReloadPageButton({ children }: Props) {
   const router = useRouter();
