+++ src/pages/locations/index.tsx
@@ -1,14 +1,14 @@
+import Container from "@/components/container";
+import SideBar from "@/components/side-bar";
+import { Post } from "@/interfaces/post";
+import { getAllPosts } from "@/lib/api";
+import { BASE_URL } from "@/lib/constants";
 import Head from "next/head";
 import Link from "next/link";
-import Container from "../../components/container";
-import SideBar from "../../components/side-bar";
-import { Post } from "../../interfaces/post";
-import { getAllPosts } from "../../lib/api";
-import { BASE_URL } from "../../lib/constants";
 
-type Props = {
+interface Props {
   allPosts: Post[];
-};
+}
 
 export default function WhereWeveBeen({ allPosts }: Props) {
   const titleText = `Locations | Best Seen on Foot`;
